refactor(auth): authMiddleware in separate file
